+++ tests/custom(argparse)/test_simple.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Tests a simple parser
 """