+++ tests/docopt/test_advanced.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Tests an advanced parser
 """
 from __future__ import annotations